requirements.txt
# Core
fastapi>=0.115.0
anthropic>=0.40.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Claude SDK
claude-agent-sdk>=0.1.0

# Telegram
python-telegram-bot>=21.0

# Redis
redis>=5.0.0

# PostgreSQL
asyncpg>=0.30.0

# HTTP
httpx>=0.27.0

# YAML
pyyaml>=6.0

# Memory pipeline
rapidfuzz>=3.0.0
google-genai>=1.0.0

# Auth
pyjwt>=2.9.0
bcrypt>=4.2.0

# Supervisor XML-RPC
supervisor>=4.2.0

# CLI interface
rich
prompt-toolkit

# Parquet (Roger/Strava)
pyarrow>=17.0.0

# Utilities
python-dotenv
tzdata
